+++ test/unit/fuzz-tools.test.js
@@ -393,6 +393,7 @@ import * as seinsheimerSubtrochV413 from '../../lib/seinsheimer-subtroch-v413.js
 import * as mayfieldPerilunateV414 from '../../lib/mayfield-perilunate-v414.js';
 import * as geisslerCarpalV415 from '../../lib/geissler-carpal-v415.js';
 import * as russeScaphoidV416 from '../../lib/russe-scaphoid-v416.js';
+import * as wasselThumbV417 from '../../lib/wassel-thumb-v417.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -731,6 +732,7 @@ const MODULES = {
   'mayfield-perilunate-v414.js': mayfieldPerilunateV414,
   'geissler-carpal-v415.js': geisslerCarpalV415,
   'russe-scaphoid-v416.js': russeScaphoidV416,
+  'wassel-thumb-v417.js': wasselThumbV417,
   'tb-testing.js': tbTesting,
 };